-register user data -Note :Add,Edit,Delete -Tag :Add,Edie,Delete -change alert dialog UI[In Progaress]

+++ app/src/main/java/com/note/compose/ui/theme/home/tag/TagScreen.kt
@@ -1,18 +1,22 @@
 package com.note.compose.ui.theme.home.tag
 
 import android.widget.Toast
+import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
+import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.text.BasicText
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Delete
 import androidx.compose.material.icons.filled.Edit
+import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.CircularProgressIndicator
@@ -20,6 +24,7 @@ import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
@@ -28,8 +33,13 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.colorResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.font.FontFamily
+import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
@@ -54,7 +64,6 @@ import com.note.compose.viewModel.TagViewModel
         }
         is ResultState.Success -> {
             tagsList = (tagState as ResultState.Success).data
-            Toast.makeText(context,"get tag Success!", Toast.LENGTH_SHORT).show()
         }
         is ResultState.Error -> {
             val error = (tagState as ResultState.Error).message
@@ -66,9 +75,10 @@ import com.note.compose.viewModel.TagViewModel
         Box(
             modifier = Modifier.fillMaxSize()  // Take the full screen size
         ) {
-            BasicText(text = "No tags yet!",
+            Text(text = stringResource(id = R.string.no_tags_yet),
                 modifier = Modifier.align(Alignment.Center),
-                style = MaterialTheme.typography.headlineMedium
+                style = MaterialTheme.typography.headlineMedium,
+                fontFamily = FontFamily.Serif
             )
         }
     }
@@ -94,24 +104,76 @@ import com.note.compose.viewModel.TagViewModel
                             text = tag.tagName,
                             modifier = Modifier
                                 .weight(1f)
-                                .align(Alignment.CenterVertically)
+                                .align(Alignment.CenterVertically),
+                            fontFamily = FontFamily.Serif,
+                            fontWeight = FontWeight.SemiBold
                         )
                         Row(
                             modifier = Modifier.align(Alignment.CenterVertically),
                             horizontalArrangement = Arrangement.End
                         ) {
+                            // Edit Button with Rounded Background
                             IconButton(onClick = {
                                 onEditTagClick(tag)
                             }) {
-                                Icon(Icons.Default.Edit, contentDescription = "Edit")
+                                Box(
+                                    modifier = Modifier
+                                        .size(32.dp) // Set the size for the circular background
+                                        .clip(CircleShape) // Make it circular
+                                        .background(color = colorResource(id = R.color.color_1A5E35B1)), // Background color
+                                    contentAlignment = Alignment.Center // Center the icon within the circle
+                                ) {
+                                    Icon(
+                                        imageVector = Icons.Default.Edit,
+                                        contentDescription = "Edit",
+                                        tint = Color.White ,// Icon color
+                                        modifier = Modifier.size(16.dp)
+                                    )
+                                }
                             }
+                            var showDialog by remember { mutableStateOf(false) } // Dialog visibility
+                            // Delete Button with Rounded Background
                             IconButton(onClick = {
-                                viewModel.deleteTag(tag.tagId)
+                                showDialog = true
                             }) {
-                                Icon(
-                                    Icons.Default.Delete,
-                                    contentDescription = "Delete",
-                                    tint = colorResource(id = R.color.color_B50202)
+                                Box(
+                                    modifier = Modifier
+                                        .size(32.dp) // Set the size for the circular background
+                                        .clip(CircleShape) // Make it circular
+                                        .background(color = colorResource(id = R.color.color_B50202)), // Background color
+                                    contentAlignment = Alignment.Center // Center the icon within the circle
+                                ) {
+                                    Icon(
+                                        imageVector = Icons.Default.Delete,
+                                        contentDescription = "Delete",
+                                        tint = Color.White ,// Icon color
+                                        modifier = Modifier.size(16.dp)
+                                    )
+                                }
+                            }
+                            // Confirmation Dialog
+                            if (showDialog) {
+                                AlertDialog(
+                                    onDismissRequest = { showDialog = false }, // Close dialog on outside click
+                                    title = {
+                                        Text(text = stringResource(id = R.string.delete_tag),fontFamily = FontFamily.Serif)
+                                    },
+                                    text = {
+                                        Text(stringResource(id = R.string.are_you_sure_delete_tag),fontFamily = FontFamily.Serif)
+                                    },
+                                    confirmButton = {
+                                        TextButton(onClick = {
+                                            showDialog = false // Close dialog
+                                            viewModel.deleteTag(tag.tagId)
+                                        }) {
+                                            Text(stringResource(id = R.string.delete), color = colorResource(id = R.color.color_B50202),fontFamily = FontFamily.Serif)
+                                        }
+                                    },
+                                    dismissButton = {
+                                        TextButton(onClick = { showDialog = false }) {
+                                            Text(stringResource(id = R.string.cancel),fontFamily = FontFamily.Serif)
+                                        }
+                                    }
                                 )
                             }
                         }
